Moved back settingRegistry dependant commands
Due to a still unclear issue, when passing settingRegistry to
functions it changes the beahvior of the commands that use it.

This change moves the six commands that interact with settingRegistry
back to index.ts from commands.ts where they were oved in #6904

This fixes #7295
